requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
colorama==0.4.6
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
factory_boy==3.3.3
Faker==40.5.1
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.16.0
pytest==9.0.2
pytest-django==4.12.0
pytest-html==4.2.0
pytest-metadata==3.1.1
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
